add interimValue param to Handler interface
Before this change, there was no way to signal to a subsequent
handler that a value is being proposed to be set as the new value.

This commit adds a new parameter `interimValue` to the `Handle`
method of the `Handler` interface.

The `interimValue` represents the current proposed value for the
Field being processed, as determined by the preceding Handlers in
the middleware chain. This allows each Handler to have access to
the current state of the value being set.

Handlers are now able to consider the `interimValue` when
processing the Field. If a Handler wants to override this value,
it can return a non-nil value. If it doesn't wish to modify the
`interimValue`, it can simply return nil. Otherwise, it's
perfectly reasonable for a handler to completely ignore the
'interimValue' passed to it, which is likely to be the case most
of the time.