only store flags that were explicitly provided to Parse.
When passing a flag with its type's zero value, the flag handler
would return <nil>, which is a bug. This commit adds the ability
to tell the difference between explicitly providing a command line
flag vs the flag returning it's default value.

The flag handler keeps a map "parsedFlags" which contains all of
the flags' values. This uses the Visit method on a flag.FlagSet
to get a map of all flags explicitly passed to Parse, then removes
all other flags in "parseFlags" that are not in this map.

When the handler goes to pull the value for the flag from the
"parsedFlags" map, it doesn't exist, and therefor now correctly
returns <nil>.

This also means that the valueFn was actually a bug, and could be
safely removed.

This is the most precise solution I could find that required the
least amount of code change, although I'm not entirely sure I like
how it's all turning out. Gonna have to let it bake for a bit.
Another option I did consider and still want to try, is creating
custom types that satisfy the flag.Value interface which would
provide a wider array of customization, like better formatted help
text, the ability to track whether it was explicitly set, and
potentially a whole host of other things. I decided against this
for now because that would be a lot more code to maintain, and
wrestles more responsibility from the stdlib flag package where
it does a perfectly acceptable job. Again, we'll let it bake.