separate functionality into different packages
The intent is to simplify the module api and provide the most
intuitive top level api while also keeping it very simple. You
shouldn't have to go read through every exported member to
understand what is happening, but I also wanted to be sure that
access to the underlying core was still available for building on
top of.

structconf
- This is the module namespace package api. It takes the
underlying functionality and abstracts a simple api for the most
common use case.

stronf
- This is the package where the core lives, and a "lower level"
api is made available to assist with building a different
abstraction or building handlers/middleware that require access to
the module's primitives.

confhandler
- This package exposes a set of premade middleware and handlers to
use however necessary.