refactor with improvements
Well, this is embarrassing. I went on and on, refactoring, and tackling
each whim I had over the course of several weeks and ended up here: with
massive changes and too lazy to logically split them all.

What started with a refactor of the core SettableFields func, led to
realizing that I didn't quite like the way the handlers were...
handling, so that changed. I found ways to split up ideas into smaller
functions, and move what was actually core logic from the structhandler
package to the stronf package. Tests were played with and so on...

Suffice it to say that this is a big change to the internals, including
breaking changes.

Hopefully I can avoid a large commit like this in the future, but what
can I say... it's my project, and it's not like any serious project
would consider using an obscure config library like this. ✌️