fix test to work with go1.21
Because I've been switching between go versions for other projects, this
one snuck by me previously and I must have shrugged it off in the last
monstor commit.

Moving to go1.21, there was an update to the regex package:

> Regexp now defines MarshalText and UnmarshalText methods. These implement
encoding.TextMarshaler and encoding.TextUnmarshaler and will be used by packages
such as encoding/json.

https://go.dev/doc/go1.21#minor_library_changes

This explains the discrepancy here, so we're just gonna be on go1.21
now.