remove RSAHandler from the module
It was a nice exercise to build it out, but ultimately I'm realizing
that it's a bit premature and presumptuous to be adding in a very
specific secret handler. I'm not even using it yet in my own projects!
I'm also finding that the way a project or organization chooses to
handle it's secrets is typically pretty bespoke, from determining how
secrets are loaded, where/how encryption/decryption is handled, their
encryption scheme, how they're stored and even rotated. There's a lot
to consider.

Either way, I think it's best to remove this footgun from the module
altogether and rethink how, or even if, a secret handler should be
implemented in a way that is considered a "sensible default".