refactor flag handler to be simpler
This is a big refactor. Of note, I've removed any reliance on generics,
which meant a pretty large set of breaking changes across structconf and
the flag handler itself, but I beleive this makes it much more simpler
to use. Previously, to use structconf, you had to initialize a top level
struct for the module, then pass a struct to be parsed. Generics were
used to assist with the flag parsing. Since I've cleaned it up and made
it capable of happening inline, we no longer needed any generics and
could remove a lot of boilerplate. Now the top level structconf package
just has a single Parse function and it's supporting optionFuncs. This
slims down the api significantly.

Other improvements to the flag handler include that it's using more of
the logic from the core stronf package, so that any changes are kept in
sync. I learned a bit more about how the stdlib flag package makes
available the generic functionality of a flag via it's Var functions.

Anyway, the removal of generics was a big motivator for this change, and
the resulting code is much simpler because of it.

I also removed the test files in the structconf package. I'm still
thinking about these as they're more like _integration_ tests with all
of the handlers, and I'm not sure that's worth doing yet.