remove Handler interface
This is something I struggled with for a long while, doing a lot of
research and having many discussions, and while I don't think that many
people would argue with the use of this interface here _and_ I think
probably more would rather have it than only the HandleFunc, I believe
removing this interface is the best move. structconf doesn't _need_ it.

This interface is pretty bespoke to this package and it's unlikely to be
any more generally useful for implementations of it outside of working
directly with this package. I've settled on an extremely high bar for
necessitating the creation of an interface, and this doesn't meet it.
Anything I originally thought I needed the Handler interface for can be
done through functions, and I believe that significantly simplifies the
entire module, especially when it comes to understanding and following
the code.