rename conf.go -> structconf.go
This is a simple change, but one that I beleive assists in readability
of the module. I consider it good practice to have a file share the
package (in this case module) name to indicate where to begin reading.

I don't do that consistently enough, but it is something I value when I
see an opportunity for it.