don't return field.Value()
In the majority of cases, this is actually okay, however, because this
module supports fields that satisfy encoding interfaces, returning the
field.Value() from a handler means that we could be potentially
returning something that cannot be converted to a []byte. For example, a
time.Time field would end up returning a struct, which is not
convertable to a []byte.

The overarching logic to consider though is that when a field is parsed,
if the handler returns nil, it doesn't attempt to do anything to the
field, there's nothing to update it with. At least, that's a part of the
philosophy of structconf: no reference types. So setting a field to nil
can only happen through the encoding interfaces. Essentially it boils
down to: the field is already set to field.Value(), no need to set it
again.

Returning nothing with no error is correct here, and rarely will
returning field.Value() be correct.