include type information when user supplied "usage" is used
The stdlib flag package has this strange way of identifying the type of
the flag by inspecting the usage information for something in backticks.
It places whatever is in backticks as the type information for the flag.
Since these usage strings are supplied through struct tags, and struct
tags won't allow arbitrary usage of backticks, this just tacks on the
type in backticks to whatever is supplied. It feels a little redundant
and hopefully a better way can be found later. Placing it in square
brackets just makes it stick out a little more from the usage.