flag handling is now optional
After looking at a few different popular config libraries, it seems the
most popular way to load configuration is through environment variables,
so that should definitely be included by default, especially because
it's not terribly complicated.

Command line flags on the other hand are significantly more complicated,
and forcing the usage of them can cause a bunch of unnecessary
headaches, as they could conflict with other flag usage or require
careful extra setup. It's best to require opt-in for the flag handler.