Allow arbitrary nesting of struct tag values.
This uses a pattern that leverages the idea that all of these values are
just maps, and an optional path can be provided to access a value at an
arbitrary nested level in the struct tags.

Why would you want to do this? I don't know, it's probably not a good
idea to go any further than one or two nestings.

However, it does make the signature a little nicer to work with as this
makes the second param optional.